# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School API",
        "description": "API for managing school data",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/api/v1/fees/items/{feeItem}": {
            "get": {
                "tags": [
                    "school-v2.4",
                    "school-v2.0"
                ],
                "summary": "Get fee item",
                "operationId": "a75b670e417c4b83b8a72b2fe98c6332",
                "parameters": [
                    {
                        "name": "feeItem",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fee item returned"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "school-v2.4",
                    "school-v2.0"
                ],
                "summary": "Update fee item",
                "operationId": "40557ced69a52bc911bf2588f532cdf0",
                "parameters": [
                    {
                        "name": "feeItem",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "school-v2.4",
                    "school-v2.0"
                ],
                "summary": "Delete fee item",
                "operationId": "b330485640498991a29fecc589470cc7",
                "parameters": [
                    {
                        "name": "feeItem",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}